requirements.txt
# RAG + 火锅顾问（Gemini + LangChain + FastAPI）
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
langchain-classic>=1.0.0
langgraph>=0.2.0
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Web 服务
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
